// packages/sdk/src/types/paymaster/paymasterEndpointDetailResponse.ts
/**
 * Kriptonio Paymaster API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import type { SponsoredTransactionResponse } from './sponsoredTransactionResponse';
import type { PaymasterResponse } from './paymasterResponse';


export interface PaymasterEndpointDetailResponse { 
    id: string;
    title: string;
    accessToken: string;
    managed: boolean;
    sponsorshipCount: number;
    wallet?: string;
    sponsorshipWei: string;
    paymaster: PaymasterResponse;
    lastSponsoredTransactions: Array<SponsoredTransactionResponse>;
    createdAt: string;
}