// packages/sdk/src/types/paymaster/sponsorTransactionResponse.ts
/**
 * Kriptonio Paymaster API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SponsorTransactionResponse { 
    paymasterAndData: string;
    preVerificationGas: string;
    verificationGasLimit: string;
    callGasLimit: string;
}